+++ praw/models/reddit/subreddit.py
@@ -3362,7 +3362,7 @@ class Subreddit(MessageableMixin, SubredditListingMixin, FullnameMixin, RedditBa
             "title": title,
             "nsfw": bool(nsfw),
             "spoiler": bool(spoiler),
-            "validate_on_submit": self._reddit.validate_on_submit,
+            "validate_on_submit": True,
         }
         for key, value in (
             ("flair_id", flair_id),
@@ -3477,7 +3477,7 @@ class Subreddit(MessageableMixin, SubredditListingMixin, FullnameMixin, RedditBa
             "spoiler": bool(spoiler),
             "sr": str(self),
             "title": title,
-            "validate_on_submit": self._reddit.validate_on_submit,
+            "validate_on_submit": True,
         }
         for key, value in (
             ("flair_id", flair_id),
@@ -3600,7 +3600,7 @@ class Subreddit(MessageableMixin, SubredditListingMixin, FullnameMixin, RedditBa
             "title": title,
             "nsfw": bool(nsfw),
             "spoiler": bool(spoiler),
-            "validate_on_submit": self._reddit.validate_on_submit,
+            "validate_on_submit": True,
         }
         for key, value in (
             ("flair_id", flair_id),
@@ -3699,7 +3699,7 @@ class Subreddit(MessageableMixin, SubredditListingMixin, FullnameMixin, RedditBa
             "title": title,
             "nsfw": bool(nsfw),
             "spoiler": bool(spoiler),
-            "validate_on_submit": self._reddit.validate_on_submit,
+            "validate_on_submit": True,
         }
         for key, value in (
             ("flair_id", flair_id),
@@ -3819,7 +3819,7 @@ class Subreddit(MessageableMixin, SubredditListingMixin, FullnameMixin, RedditBa
             "title": title,
             "nsfw": bool(nsfw),
             "spoiler": bool(spoiler),
-            "validate_on_submit": self._reddit.validate_on_submit,
+            "validate_on_submit": True,
         }
         for key, value in (
             ("flair_id", flair_id),