+++ CHANGES.rst
@@ -10,6 +10,7 @@ Unreleased
 
 - Remove ``APIException`` class.
 - Remove ``WebSocketException.original_exception`` method.
+- Remove ``Reddit.validate_on_submit`` configuration attribute.
 
 7.8.1 (2024/10/25)
 ------------------